backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
celery==5.3.6
redis==5.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
PyGithub==2.2.0
django-environ==0.11.2
gunicorn==22.0.0
whitenoise==6.7.0
